Develop

+++ db/migrate/20210907085627_create_post_comments.rb
@@ -2,8 +2,8 @@ class CreatePostComments < ActiveRecord::Migration[5.2]
   def change
     create_table :post_comments do |t|
       t.text :comment
-      t.integer :user_id
-      t.integer :post_id
+      t.bigint :user_id
+      t.bigint :post_id
 
       t.timestamps
     end
